% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes_in.R
\name{primes_in}
\alias{primes_in}
\title{Finds prime numbers in a range of integers.}
\usage{
primes_in(x, y)
}
\arguments{
\item{X}{a positive integer (or 0)}

\item{Y}{a positive integer, higher than or equal to X.}
}
\value{
a list of the primes that are found between (and including) numbers X to Y.
}
\description{
Finds prime numbers in a range of integers.
}
\examples{
primes_in(1, 20)
primes_in(5, 40)
}

